+++ packages/flame/lib/src/components/core/component.dart
