+++ packages/flame/lib/src/components/mixins/coordinate_transform.dart
@@ -1,4 +1,4 @@
-import 'package:flame/src/components/component.dart';
+import 'package:flame/src/components/core/component.dart';
 import 'package:vector_math/vector_math_64.dart';
 
 /// Interface to be implemented by components that perform a coordinate change.
