+++ packages/flame/lib/src/components/router_component.dart
@@ -1,4 +1,4 @@
-import 'package:flame/src/components/component.dart';
+import 'package:flame/src/components/core/component.dart';
 import 'package:flame/src/components/overlay_route.dart';
 import 'package:flame/src/components/route.dart';
 import 'package:flame/src/components/value_route.dart';
