+++ packages/flame/lib/src/components/sprite_batch_component.dart
@@ -1,6 +1,6 @@
 import 'dart:ui';
 
-import 'package:flame/src/components/component.dart';
+import 'package:flame/src/components/core/component.dart';
 import 'package:flame/src/sprite_batch.dart';
 import 'package:meta/meta.dart';
