+++ packages/flame/lib/src/events/messages/position_event.dart
@@ -1,5 +1,5 @@
 import 'package:flame/experimental.dart';
-import 'package:flame/src/components/component.dart';
+import 'package:flame/src/components/core/component.dart';
 import 'package:flame/src/events/messages/event.dart';
 import 'package:meta/meta.dart';
 import 'package:vector_math/vector_math_64.dart';
