+++ packages/flame/lib/src/game/mixins/single_game_instance.dart
@@ -1,4 +1,4 @@
-import 'package:flame/src/components/component.dart';
+import 'package:flame/src/components/core/component.dart';
 import 'package:flame/src/game/game.dart';
 
 /// Mixin that declares a [Game] class as a singleton.
